Remove storages from database in separate loop
Before this commit, removing storages from the database was
in the same loop as going through the database to find all those
databases. This can cause various iteration issues, for example
the loop can skip entries that should be removed as well.
This commit ensures that the storages are removed after whole model
was searched for storages to remove, so no skipped entries can occur.

Probable steps to reproduce the issue before this commit:
- have three bundles called very similarly, like Bundle1, Bundle2,
Bundle3
- remove all three and run Krita again
- see that Bundle2 is still there but you can't access the resources;
Krita complains in the console that the storage is not found