Fix saving to bundle
Before this commit, bundle saving code would operate only on
filenames, but the dlg_create_bundle code changed that filename
to be something else. So the bundle would save with different filename
in the manifest and different actual filename.
This commit introduces two new parameters to the KoResourceManifest
(no new things are added to the actual manifest file in a bundle)
that are resourceId and filenameInBundle.

The first one ensures the first resource the bundle will be looking for
is the one with the same resouce id, if it's available.
The second one helps sanitizing the filename in the newly created bundle
while still allowing the manifest to point to the correct file.

BUG:443907