Fix Discrete Canvas Rotation sometimes skipping steps
Before this commit the Discrete Canvas Rotation mode would
be unable to snap to the original rotation if starting
from an angle that was a multiple of 15. This was due to the
mode trying to ignore snapping to that angle at the start
of the action, but was applied through the whole action.
This commit fixes that issue.

BUG:429637
(cherry picked from commit de158f5e538cb40ec37eba00670a8be3567d066b)