Fix overwrite dialog in seexpr looking like create dialog
This commit disables the name in the overwrite seexpr dialog, because
even if it is enabled, the name change doesn't last; moreover
it looks nearly the same as create new script dialog.
Disabled name textbox makes it look more like you're saving
a new version (maybe with a new thumbnail, but still the same thing).