Fix crash on loading `tmp.kts copy` as a taskset
Before this commit, if one tried to use native file dialogs
and import a file called for example `tmp.kts copy` as a taskset,
Krita would crash or assert.
This commit ensures that Krita recognizes that the loader hasn't been
created and leaves the function early, which avoids the crash.

CCBUG:444950