Fix issues when soft proofing is active during colorspace conversion
The color transform needs to be reset before updating the canvas.
Also, there was an odd workaround to set new channel flags for the root
layer, but this was an async signal that caused an update in the middle
of invalidating the image data during conversion.
Instead, simply fetch the (after the previous fix valid) new channel flags
when the canvas gets notified about the color space change.

BUG:441165
BUG:426641
CCBUG:329336