Make sure that group layers can still have non-image color space
Test case:

0) Create RGB8 image
1) Create two paint layers and put *one** of them into a group
2) Convert group's color space to GrayA8
3) Confirm that the first layer looks gray now
4) Move the second paint layer (which is stll RGB8) into a group
5) This layer should also become looking gray, because of the group's
   color space.

From the technical point of view the patch removes `convertImage`
parameter and makes sure that the color space of the group layer
is always reset to the color space of the destination conversion.
Before the MR, this was done by conversion functions in
visitNodeWithPaintDevice() and was lost due to too simplified nature of
KisDoSomethingCommand, which always resets to image's color space.