Fix caching bug when scrubbing from cached to uncached frame.
An optimization to prevent unnecessary refreshes was causing a
regression. Since scrubbing on cached frames doesn't trigger a
time change, but uncached frames do, an optimization where
we check if the last frame and the current frame exist within
the same "timeframe" would lock users out from seeing the
uncached frames properly when scrubbing.

We now force at least one time switch call when switching from
uncached to cached frames to prevent this issue. Now, it will
properly recognize that the last frame and the 'current frame'
are different and refresh their contents.

BUG:445265