Fix embedding of the linked resources into .kra file
KisDocument should **not** return raw resources in
linkedDocumentResources(), because when it reads from the storage
directly, the resource become not fully initialized. Its resourceId(),
storageLocation(), active() and md5() fields are not initialized.
These fields are managed by KisResourceLocator, but here we short
circuit by reading from the storage manually. Therefore we can only
manipulate the objects of KoEmbeddedResource, that is, serialized
blobs.