Refactor-out the usage of QBuffer with dummy QByteArray
In quite a lot of places we use outdated QBuffer's API, where we
create a QByteArray jsut for the sake of passing it into QBuffer.
The current QBuffer's API creates an internal byte array internally
in the default constructor.

(That is not something life-critical, I just felt itchy when saw
that everywhere =) )