Fix edit, rename and cancel operations on gradients
We cleaned up the function to work this way:

1. If the gradient is in edit mode (i.e editGradient = true). We save
the gradient right away, even if it is renaming and most importantly we
do not create duplicate gradient with the same name (which was
previously caused by oldname == newName check).

2. If we're adding a newly created gradient, we do all the checks like
before to ask user if they want to overwrite a resource with the same
filename.

3. At all times we operate on copy of the original gradient, this
prevents us from updating the gradient when a cancel operation is used.

(Thanks to tiar for the review)

# Conflicts:
#	libs/ui/widgets/gradient/KisGradientChooser.cpp