Improve font style selection, enabling the proper styles to be selected.
Apparantly Qt will synthesize the bold/italic if you don't set it on creation.
This patch also tries to fix the handling of stretch in the same way, however
it seems that all QFonts return 'anystretch' for the stretch, which makes me
suspect there's a qt bug there. In any case, the changes ensure that at the
least a coherent value is set (0 doesn't mean much).

BUG:425312