Change null strings to empty in CacheDb queries
Before this commit, CacheDb queries could sometimes
fail because the string supplied was null instead of empty
(most common in storages, where an empty string indicates
the resource folder).
Null strings in queries cause issues because comparison with a null
is always string.

This commit ensures that null strings are converted to empty strings
before being used in in WHERE part in an sql query.

CCBUG:444975