Fix random memory corruptions when manipulating layers too quickly
It turned out, when the user switches visibility of a layer,
the event loop can be entered recursively (due to KisBusyWaitBroker),
it could lead to memory corruptions due to the broken pointers
to KisNodeDummy stored in QModelIndex objects.

The problem is fixed by forcefully syncing the node manipulation
signals to non-recursive event processing using a new
KisSynchronizedConnection class. See documentation in it for details.

BUG:444516