[not for 5.0] Fix flickering of tilt-controlled brushes on tablet release
When the stylus leaves the tablet surface a pair of
KisToolPaint::activatePrimaryAction()/deactivatePrimaryAction() events
is delivered to the tool. They request update of the tool outline, but
since there is no pointer event available, this outline was generated
via a non-tilted/non-rotated fallback case. It caused a visible flickering
while painting.

The patch implements multiple things:

1) It severely refactors KoPointerEvent to allow copying and storage
   of such events. Now the class became "much more simple and
   understandable" (c) due to the usage of boost::variant2 (which
   is technically a C++17's std::variant).

2) Now the tool can request the last delivered pointer event from the
   canvas via KoToolBase::lastDeliveredPointerEvent().

BUG:445476