Fix wrong resource selection on creation and import
Before this commit, the resource selected in the palette chooser
wasn't always the same as the palette shown in the editor.
This could cause the user to delete a different palette that they wanted
if they didn't notice. It could also cause the palette to lose all the
added colors when switching to the same palette.