Make hashing function in the tile manager more robust
In same rare cases (which are officially bugs and must be fixed), the
algorithms may request tiles from the tile data manager from
non-addressable range, which is above +/- 2 million pixels. The tile
manager cannot address anything there, so before the patch it just
crashed with an assert. This patch add a fallback solution: the such
a case, the tile manager will just return an empty detached tile to
the caller. This tile will be destroyed via a shared pointer when
the caller is done processing it.

CCBUG:441826