Change safe assert to a console warning
Before this commit, there was a safe assert every time a preset
with a brush tip from ABR file was created or if the user tried to make
a bundle with such a preset.
This commit changes the safe assert to a console warning.
Ultimately it should be made possible to just embed the brush tip
in the preset regardless if it's serializable or not (if it can be
converted into a serializable).