Fix regression in positioning of URL-based drag-and-drops
47db9c0be50fc85f38ba6c86757a7b307f0bbbd4 did not account for the
position of the drop event, and instead used the cursor position.
This means that, if the menu and/or the dialog popup were invoked, the
paste position would be taken from the interaction with those widgets
and not from the place where the action was triggered.

This doesn't fix the discrepancy in painting a remote and a local
image; that will be fixed in the next commit.

CCBUG: 446120