Fixed backwards compatibility edge case for alpha-mask png brush tips.
User was getting unexpected results compared to previous version of krita.
While this behavior is not technically a bug, it was a regression compared
to the user's expected behavior. They had a single png image with all black
rgb channels and only an alpha channel defined. We can try to catch this
specific edge case to assign the alpha mask mode to the correct setting
to have consistent behavior with krita 4...

BUG:445691