Fix transform tool rotation behaving erratically
Before this commit if a rotation with the transform
tool was started below the pivot point the transform
would rotate in the opposite direction of the cursor
movement at certain points in the rotation. This was
due to the transform tool not having the correct
clockwise direction from the starting angle in
those cases.
This commit fixes that by fixing how the clockwise
direction is determined.