Fix line height being too huge in Krita 5.0
All the internals of QTextDocument work with the primary screen's DPI.
That is, when we ask Qt about font metrics, it returns values scaled
to the pixels in the current screen, that is,
ptValue * qt_defaultDpi() / 72.0. We need to convert this value back
into points before writing as SVG. Therefore, all the metrics returned
by the document are scaled with `fixQtDpi()`.

Official Qt's way to workaround this problem is to set logicalDpiX/Y()
values on the paint device's associated with the document. But it seems
like in our version of Qt (5.12.12, which is rather old) it doesn't work
properly.

BUG:443795