Add unit test for two resource models
Before this commit, there was no unit tests that would check if
one resource model would see changes done in another resource model
if they were both created before the change happened.
This commit adds a unit test that checks this.

Note: at the time of writing, the test works fine if it's the only
test being run. But it fails if other tests were run before too.

Partial conclusions: the reason for the test failing is mismatch
in indexing, probably caused by addTemporaryResources() test
that adds a second resource with the same filename (even though
supposedly in a different storage). Possible guilty party: GROUPBY
statement in KisResourcesModel. If the filename is changed
(so all of them are unique), there are no issues and the test
passes successfully.