Fix embedded resources to appear in the resource model
All the code that manipuilates resources in Krita is expected to do that
via KisResourceModel. It guarantees that the internal model indexes are
valid and up-to-date.

Though we have one notable exception, that is, embedded resources. They
are loaded by KisResourceLocator internally, when some uncached resource
is loaded. In such a case the model should be notified to update its
query.

BUG:441350