Another pass at alpha mask related regression.
It seems like the behavior changed due to an unnecessary
change to a default value in the KisColorfulBrush header.
Changing the autoadjust bool to start at true was causing
unintended modifications to ALPHAMASK brush behavior.

Checked unit tests and all tests passed after changing the
default value back -- plus the behavior was fixed and did
not seem to cause regressions anywhere else.

BUG:445691