Fix an assert when loading a layer style from Krita 4.x days
The filename of a resource in Krita cannot have slashes inside,
because it confuses storages. And gradients created in Krita 4.x
and embedded into ASL could have full path in their filename.
In this particular case we can just safely convert the filename,
since it is not used for addressing in ASL.

BUG:446601