[not for 5.0] Add a sanity check for KoResource filename
The resource filenames in Krita 5.x cannot have slashes inside, that is,
they should be "file names", not "file paths". That is quite a dramatic
change that happened during 4.x->5.x transition. And some of our code
that loads the embedded resources can still initialize the resource with
the full file path.

This patch adds a sanity check into KoResource to catch this case earlier.

CCBUG:446601