Make tags dynamically translatable
This adds a new table to the db scheme, removes the kconfig* backend
for reading ini files in favour of a hand-coded parser which can give
us all translated texts and handles getting the translated texts in
the tag model.

Note that this completely breaks saving tags from older database
schema's, so good thing we're still in beta.

BUG:439300