Implement toolbar button highlight color change differently
Reacting to mouse button events missed the case when toggling buttons
via shortcut.
There is just no event that directly correlates with button toggles
other than the triggered paint event that happens on other occasions too,
so using eventFilter() for that is more of a hack.
This commit moves it to a signal handler instead.

Thanks Tom Tom Tom for pointing out the issue.