Fix regression where cursor would stay hidden after closing popup palette on certain systems.
Users would right click with mouse and then close with tablet pen, which on certain QT systems
would cause the cursor to stay hidden. I guess on newer versions of QT, not "accepting" an input
is somehow corrected or acommodated for when it comes to cursor change operations.

Adding a "setAccepted" call after hiding an open popup palette seemed to fix this issue.

BUG:441370