Add QScroller support for WGColorPatches
Also move the additional buttons to the lower right border.
Kinetic scrolling is not setup automatically yet, the existing
functions only support QAbstractScrollArea.
Although this class is now structured much like QAbstractScrollArea
and maybe it would've been easier to use base it on that, but would
require refactoring due to the QObject single inheritance limitation.