Remove variantFromResourceQueryByID from the resource query mapper
The mapper can now map both the KisResourceModel query and the
KisTagResourceModel query to variants, so we don't have to query
by id anymore for KistagResourceModel.

This fixes 446984, but it's still weird that this was fine in 2020,
and started making creating images slow end of 2021. In any case,
creating images is cheap again.

BUG:446984
(cherry picked from commit a3272dd378e3b622ebab450f606ebab2a1470eec)