Workaround for removing tag when searching for tag in text field.
It's a bit hacky, but there's a work around now for updating the search
results to match changes made to tags. It seemed more logical to track
for tag removal and additions within the ProxyModel itself, but for
some reason invalidateFilter would only update after removing
a second resource from the same tag.

BUG:446647