Add an "active" column to the resource_tags table
* This is a db schema change, but since we're simply adding a column,
I've implemented an update table statement for this, so the database
isn't dropped and recreated. If you want to test the database update,
make sure you keep a copy of the current version around before starting
a build of this MR.

* untagging a resource now sets the resource_tag entry's active column
to 0. This means there are three states for a resource: never tagged,
tagged but untagged and still tagged. resource_tag entries are now not
deleted from the database if a resource is untagged.

BUG:447134
(cherry picked from commit 791aeda4f350c8f012dde5da4696590481719a02)