Fix an ASAN crash when switching to "g) Dry Bristles" from any RGBA brush
The KisPredefinedBrushChooser uses the combo box to select the mode,
so it should be consistent with what the actual brush supports.

We need a better fix to make KisColorfulBrush more robust and
actually convert brush tip image to RGBA32 format automatically.
Right not the crash happens with grayscale brushes that use Indexed8
format.

BUG:447131