Fix artifacts when using colorsmudge brush over CMYK image
The problem is that we cannot rely on LCMS' conversions between
CMYK-U8 and CMYK-U16. They cannot do proper round-tripping, the
color fluctuates a lot, therefore causing artifacts.

To workaround this issue we can just offload all these conversions
to KoOptimizedRgbPixelDataScalerU8ToU16 (which now supports varying
number of channels in the pixel).

BUG:447211