Android: Work around crash due to Qt Accessibility
Inside Qt:

The problem is more complex than anticipated. We get to this function
from the Android UI thread and we're supposed to return the node.
However accessing QWidgets from non qtMainLoopThread is
dangerous (results in a crash).

A solution that comes to mind is to invokeMethod while blocking, but
that results in deadlocks (because qtMainLoopThread also has to wait for
Android Ui thread).

NOTE: Async solution has its own complications:
1. We'd have to manage lifetime of jobject local refs
2. The node behind the scene may be long gone.