Rewrite KoDockWidgetTitleBar to use QHBoxLayout
Instead of custom-painting the title text and positioning the buttons
manually, just use a label for the title and use a QHBoxLayout to
position everything. This has these advantages:

- Title doesn't overlap with buttons anymore
- Title can be elided by ellipsis, and the full title can be shown in a
  tooltip
- We can allow the docker to become narrower without breaking the
  visuals too much
- The button layout is no longer broken in RTL

The sizeHint/minimumSizeHint calculation has been removed in favour of
letting the layout handle it automatically. We still do funky stuff to
calculate the button size, but this shouldn't be a huge problem.