Update Image Properties to display correct # of layers
Previously the layer count shown to the user was inflated by one, since all of the layers the user created are attached to the rootLayer which was also being counted.

This creates confusion for the user as a backend only node was calculated into what they can see exists.

To get the count we need, a new helper function that counts all dependent layer nodes of the rootLayer.

The nlayers function has not been modified, since it's result is expected in the backend and depended on for unit tests and other functionality.

This allows us to show a valid user facing value without adjusting the internal logic.