Replace double slashes in recent file paths with single slash
Sometimes KConfigGroup::readPathEntry substitutes `$HOME` with the
actual home path and leaves a double slash in the result. This not only
looks bad to the user, but it also affects the ability of
KRecentFilesAction to deduplicate the paths and prevents
KisRecentDocumentsModelWrapper from correctly caching the file thumbnail
icon, and in some circumstances multiplies the workload.

Removing the extra slash fixes these issues.