macOS add support for notarytool
We notarize using altool, but its been deprecated since 2021. Notarytool
uses different servers, upload and notarization are faster and we do not
need to probe the status periodically to know if it finished.

It will be used automatically if signature, account and password are used
It is adviced to register credentials to notarytool like this:

xcrun notarytool store-credentials --apple-id <apple-id> --password <app-pass>
 --team-id <team-id> KritaNotarizeAccount

In which case only the signature is needed for further notarization calls

note: team-id can be checked from appleId -> membership, but as it is
also part of the certificates if not provided the script will try to
get it from the last 10 digits of the signature.

Finally, this ports osxdeploy.sh to be executed by zsh

(cherry picked from commit 89b1bc1624dbf2be43fbb7d83f5bc7360d74c396)