Fix a crash when switching to a snapshot with fill layers
We should replace the root before amitting any signals, because
some of the layers may be subscribed to sigSizeChanged() signal
(e.g. KisSelectionBasedLayer). So the old layers should be fully
detached before we actually emit this signal.

We should also explicitly detach the old root from the image to
make sure it doesn't react on the new signals.

BUG:447599