Correctly locate the place the user has selected for app data
This centralizes the locating of the configurable resources dir,
which apart from real resources also contains stuff like color
schemes, profiles and so on.

We don't install all of that locally though, that's only done
with real resources.

KoResourcePaths now adds the configured folder _to_ the list of
default locations, so all attempts to locate local and installed
stuff works out.

This also adds a getAppDataLocation to the libkis API so python
plugins can use it to locate, e.g., the right place to install
plugins.

BUG:447235
BUG:447813

Test plan:

1. Change the default resource location
2. Check functionality:
* install a python plugin from a url and a local one
* check whether the layer styles are saved in the right place
* check whether icc profiles are found, and whether an icc profile placed
in the resource folder is found
* ditto with color schemes
* ditto with input profiles