Fix action text being untranslated in Configure Toolbar
In the "Current actions" list we should use QAction::text() like the
"Available list", instead of QAction::iconText(), because the `iconText`
defined in .action files are not being extracted for i18n and therefore
can be untranslated when different from `text`.

Maybe we should review how and where `iconText` is actually being used,
and either remove it or make sure they are translated?

BUG: 448027