Add a safety measure on copying a resource file
Before this commit, when importing a resource file to the folder
storage, Krita would only check if the result file can be opened
for writing, not if the writing was successful.
This commit ensures that if the file was only half-written, it will
still return false, indicating a failure.

(cherry picked from commit c5b89e89a565ceeb761ac80b629861a0d3fd187d)