Fix and extend some shortcut actions for popups
The "Brushes and Stuff" actions for showing the gradients, patterns
and workspace popup were not actually connected to anything, and hence
wouldn't do anything.

Also, those actions, plus the preset chooser and the brush editor action
are now added to the respective popup widget so they can be closed again
with the shortcut used to show them.