Revert "Make recent files list show filename only once"
This reverts commit 4c3d3ccb1c9b46ae7f2afc648d45c222dc16af68.

The implementation is incorrect for Windows, that the path ends up using
forward slash instead of backward slash, and if the file is located on
the root of a drive it becomes something like `C:/\`. I also have
concern that it may fail for non-local paths, although I have not
tested any.

I may consider reimplementing it when rewriting more parts of the recent
file list handling.