Fix brush outline when the tool is switched by the new stylus proximity
Normally, when we switch tool we switch it either by a button press
or a mouse click. Both these actions call prepareReadyShortcuts()+
tryActivateReadyShortcut() as a normal routine. But when a new tool
is activated implicitly by presenting a new stylus to Krita, there was
no activation, so the tool's primary action was not activated and,
therefore, there was no outline.

BUG:442342