Don't leave incorrect resource files after loading
Before this commit, KisFolderStorage wouldn't check if the
resource can be loaded after importing, so it didn't even know
that maybe it imported an incorrect file which is later
staying in the resource folder unnecessarily.
This commit adds checking for whether the resource can be loaded
immediately after copying it to the folder, and then removes it
if the resource file is incorrect.

BUG:446279
(cherry picked from commit c76519a76452f6d7cdd2eba0911861f3ad88dd6e)