Fix wrong color smudge radius range
Before this commit, if you:
1) took a brush like `i) Wet Paint Details`,
2) went to Smudge Length and enabled New Algorithm for smudging,
3) went to Smudge Radius (the slider range would be up to 100% because
that's the range for the new algorithm)
4) reset the brush
the range of the Smudge Radius slider would be still up to 100%
even though the old algorithm (used for that brush) is up to 300%
and should be restored when resetting.

This commit fixes this behaviour and now the range matches the
algorithm.

BUG:441682