Fix sizing of preview images
There are inconsistencies in the resizing when the image is either too
small or the height is bigger than the width.

Applying the centering correction against maxIconSize will cause parts
of the image to be chopped off if the iconSize is smaller than the
result size.

Also draw preview in high res on Hi-DPI displays.

BUG: 447805
(cherry picked from commit 7770a821e6f2cb66c45da14ca3ade4db39914167)