Reset the assistant after each stroke
Before this commit, if the user created multiple assistants,
made some strokes when having snap to assistante enabled,
and then disabled all of the assistants using the assistant tool,
there would be still snapping enabled to the last used assistant.
This commit ensures that the assistant is reset in such cases.

BUG:448187