Improve performance of query in KisTagResourceModel
Before this commit, the sql query in KisTagResourceModel
was simple but had two issues:
- used GROUP BY in a potentially incorrect way (there was no way
to determine which rows will be selected, they could be different
every time; in some sql databases, this query wouldn't be accepted
as valid)
- had performance issues
This commit changes the query to be more correct and faster.
The resulting rows will be always the same because it always takes
the row with the smallest resource_id (from the grouped rows).

Other than the choice of values for the grouped rows,
this query should return exactly the same rows as the old query.

However note that this commit adds ORDER BY,
so the order of the rows will be known, the rows will be ordered
by the `id` value in the table `resource_tags`.
(That order is easiest to manage in the model when inserting and
deleting).

(cherry picked from commit aa6b042f515b0e60d8fa024112f6bd02eb305e97)