win: Add info on redirected AppData dir when running packaged
The packaged Desktop Bridge app redirects certain writes to the AppData
directories to a private app location, which makes "Open Resources
Folder" fail to work normally in some cases. This tries to detect if
such private app location is being used and informs the user about it,
also allows the user to choose to open the user AppData location or the
private app location.

BUG: 448472