Remove ad-hoc multithreading from KisAutoBrush
This multithreading has been implemented when we didn't have full AVX
support for the brushes, so it could give some benefits. For the current
implementation it seems to make the colorsmudge brush just slower. Too
much time is spent on sychronizing the threads. Especially, with CPUs
with high number of threads.