Use glBlend to replace needing FBO for tool outline
The previous behaivour requires rendering the canvas with an FBO, then
uses its texture as an input to a fragment shader to blend the tool
outline. The arithmentic:

    outColor = srcColor * vec4(vec3(1.0 - dstColor), 1.0)

... is equivalent to:

    outColor = srcColor * 1.0 - srcColor * dstColor

... except for the alpha channel, which is effectively 1.0.

This can be replicated by GL_BLEND using glBlendFuncSeparate and
glBlendEquationSeparate. This avoids needing to sample from the FBO
texture. It also works on OpenGL 2.1, so we can remove the old
`glLogicOp(GL_XOR)` codepath to get the same behaviour across all
supported systems.