Add unit tests for signals in KisTagResourceModel
Before this commit, there was no way to check whether the signals
in KisTagResourceModel are correct/expected without doing manual
testing in the Resource Manager (and even then, it was never
clear if the signals are really always correct or not).
This commit adds unit tests checking for the correctness
of signals sent by the KisTagResourceModel.

In the future, the ModelSignalChecker class can be used for
other models to ensure their signal correctness.

Note that when having multiple beginInsert... and beginRemove...
signals, they can end up in the checker in a different
(opposite) order than they are called in the model
(note that our models don't emit the signals directly but only
use the functions begin...(); and the Qt underlying
I observed it with tagResources() and I'm not sure yet
of the reason why that happens. The documentation of those signals
is not detailed enough for complex scenerios (like sending multiple
beginInsert... or beginRemove... signals).

---

Note that when having multiple beginInsert... and beginRemove...
calls, the rowsAboutToBeInserted and rowsAboutToBeRemoved
signals can end up in the checked in a different (opposite) order.
I observed it specifically in testing tagResources()
(with beginInsert... call), and I assume the reason for it
are internal indices row calculations within the
QAbstractItemModel class code.

The documentation of those signals and calls is not detailed
enough for complex scenerios (like calling beginInsert... or
beginRemove... multiple times). (Hence explanations in comments
in tagResources() and untagResources() functions).

(cherry picked from commit 9f19205a421efecf37e24a500978ec29591c2643)