Properly handle intermediate touch state
Prior to commit: If a three finger gesture is being made, because of the
normal delay the third touchpoint may not be immediately sent along with
other two touch points. This results in the state-machine first choosing
the action matching two touchpoints and then three touch points,
effectively zeroing out the effect in some cases (e.g undo then redo).

Now: For the first few events we don't immediately act on the event. But
keep track of an event with highest touch point count. Then if the same
number of touchpoints or greater are active past the first few ones, we
run the configured action. If not, the bestCandidateTouchEvent (one with
highest touch points) is fired once.