Fix saving tags on startup without db migration plan
Before this commit, if there was no migration plan
between the version of the database the user has and
the version of the database current Krita has,
Krita would attempt to save tags.
It wouldn't be successful because ResourceLocator::
saveTags() function was dependant on d->resourceLocation
variable, which wasn't initialized yet
because ResourceLocator is initialized after CacheDb
is initialized (and saveTags() is needed during initialization
of ResourceCacheDb).

This commit ensures that KisResourceLocator::saveTags()
function is static and doesn't depend on initialization
of KisResourceLocator itself, so can be used during
initialization of ResourceCacheDb, and can actually save
tags in the correct place.