Add ORDER BY to KisTagModel query
Before this commit, there was no ORDER BY in KisTagModel
main query. ORDER BY is necessary if any specific order is wanted.
In KisTagModel, the best column to order by is tags.id
because then beginInsertRows indices are easy to calculate
(a new row when adding a new tag will always happen at the end
of the model).