Fix TestKoProgressUpdater
The patch also fixes a bug caught by the unittest: if KoProgressUpdater
is created in the finished state, then it automatically deletes the
parent KoUpdater, which is not an expected behavior.

The patch changes that, KoProgressUpdater is created in somewhat started,
but incomplete state, but KoViewManager explicitly completes it in the
constructor.