Wrap Recorder Export into QScopedPointer
Before this commit, Coverity complained about
::Private::ui pointer going out of scope without being freed.
It doesn't seem to be the case in the code
because it seems to be freed by the owner,
but it's a good idea to wrap it into QScopedPointer.
Private class instance itself seems to lead to a memory leak
so this commit wraps the pointer to Private into a QScopedPointer
as well.

CID 329462