Free memory of ColorConverter in Seexpr generator
Before this commit, the memory that is allocated by the
KoColorSpaceRegistry to create a ColorConverter is never
freed (at least it seems so from the code).
This commit ensures the memroy is freed just afterwards.

Note that other generators like posterize and simplex noise
also free the memory returned by those functions, so it doesn't
seem to be cached, so it indeed belongs to the code
that called that function.

CID 329469