Fix checking for preset version
Before this commit, the code would first check the version
of the preset, if it's equal to 2.2 or not, and then
evaluate string "5.0" as bool. This is constant and always
evaluated as true. So the condition wasn't working.
This commit restores it to the original intention.

CID 350215
CID 350211


(cherry picked from commit 03ea660a31649792b594c139e658d09f24e4544c)