Avoid overflow in KoImageDataPrivate::generateKey
Before this commit, an overflow could happen because
the x could be any number up to 7, and 8*7 = 56,
which means that the operator << might like to shift
the number even 56 bits to the left, but it's all happening
on int which can have different lengths, and often is only
32 bits, which would lead to overflow.
After that operation the result is casted to qint64.
It makes more sense to first cast to qint64, and then
shift the bits.

CID 248898


(cherry picked from commit 9ab23547e9ac3649e273e76b2a4129af6b505555)